+++ CHANGES.rst
@@ -17,6 +17,7 @@ in progress
 - [ntfy] Add dedicated service plugin ``ntfy``
 - [ntfy] Use RFC 2047 for encoding HTTP header values
 - [ntfy] Add more fields: icon, cache, firebase, unifiedpush
+- [ntfy] Improve example/tutorial about Frigate event notifications
 
 
 2023-04-11 0.33.0